chore/#23: 리팩토링 및 정리
기능 테스트는 사용자 세션이 어느정도 정리되면 마저 진행하겠습니다
